queries and command all in one
