--- Dayana/Shared/Persistence/Models/Identity/Base/Auth/TokenResult.cs
@@ -1,6 +1,0 @@
-﻿namespace Dayana.Shared.Persistence.Models.Identity.Results.Auth;
-
-public class TokenResult
-{
-    public string AccessToken { get; set; }
-}
