+++ Dayana/Shared/Persistence/Models/Identity/Commands/AuthCommandsFile.cs
@@ -1,8 +1,13 @@
 ﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
 using Dayana.Shared.Infrastructure.Operations;
 using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Dayana.Shared.Persistence.Models.Identity.Commands.Auth;
+namespace Dayana.Shared.Persistence.Models.Identity.Commands;
 
 public class LoginCommand : IRequestInfo, IRequest<OperationResult>
 {
