--- Dayana/Shared/Persistence/Models/Identity/Commands/Users/DeleteUserCommand.cs
@@ -1,16 +1,0 @@
-﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
-using Dayana.Shared.Infrastructure.Operations;
-using MediatR;
-
-namespace Dayana.Shared.Persistence.Models.Identity.Commands.Users;
-
-public class DeleteUserCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public DeleteUserCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-
-    public int UserId { get; set; }
-    public RequestInfo RequestInfo { get; private set; }
-}
